backend/api_dj/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
distlib==0.3.1
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-serializer-extensions==2.0.1
djangorestframework-simplejwt==4.7.0
filelock==3.0.12
hashids==1.3.1
psycopg2==2.8.6
PyJWT==2.1.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
virtualenv==20.4.3
